Add tests to the endpoints
Add several tests to test the endpoints. It should already cover
the code decently (more than 80%). Also update some logic to
fit the test more, such as when get a single ride it does not make
sense for the endpoint to return an array of rides, since the
expectation is to return a single ride anyway.